# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Soleil Hostel API
  version: 2.0.0
  description: |
    Production-ready hostel management API for Soleil Hostel.

    ## Features
    - **Dual Auth Modes**: Bearer Token (stateless) + HttpOnly Cookie (web browsers)
    - **Unified Auth Endpoints**: Mode-agnostic endpoints that auto-detect auth method
    - **RBAC**: Three roles — USER, MODERATOR, ADMIN
    - **Optimistic Locking**: Room updates require `lock_version` to prevent lost updates
    - **Pessimistic Locking**: Booking creation uses `SELECT ... FOR UPDATE` to prevent double-booking
    - **Rate Limiting**: Multi-tier rate limiting (IP, user, endpoint-specific)
    - **Correlation ID**: Include `X-Correlation-ID` header for request tracing

    ## Authentication Quick Start

    **Option A — Bearer Token (mobile / SPA):**
    ```
    POST /api/auth/login-v2  →  { "token": "abc123..." }
    Then: Authorization: Bearer abc123...
    ```

    **Option B — HttpOnly Cookie (web browser):**
    ```
    POST /api/auth/login-httponly  →  Set-Cookie: soleil_token=...
    Cookies sent automatically on subsequent requests.
    ```

    **Option C — Unified Endpoints (auto-detect):**
    ```
    GET /api/auth/unified/me  →  Works with either auth mode
    ```

    ## Rate Limits
    - Login: 5 requests/min per IP
    - General API (public): 200 requests/min
    - General API (authenticated): 500 requests/min
    - Booking creation: 3 requests/min per user

    ## Deprecation Policy
...
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: |
        Sanctum bearer token for stateless API clients (mobile apps, SPAs).
        Obtained via `POST /auth/login-v2`.
        Include in header: `Authorization: Bearer {token}`

    CookieAuth:
      type: apiKey
      in: cookie
      name: soleil_token
      description: |
        HttpOnly cookie authentication for web browsers.
        Login via `POST /auth/login-httponly`, then cookie is auto-sent.
        CSRF protection via SameSite=Strict.
  schemas:
...
    Room:
      type: object
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: "Ocean View Double"
        description:
          type: string
          nullable: true
          example: "Spacious room with ocean view"
        price:
          type: number
          format: decimal
          example: 120.00
        max_guests:
          type: integer
          minimum: 1
          maximum: 10
          example: 2
        status:
          type: string
          enum: [available, unavailable]
          example: "available"
        lock_version:
          type: integer
          description: |
            Optimistic locking version. MUST be included in PUT/PATCH requests.
            Backend returns 409 Conflict if version doesn't match.
          example: 5
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Booking:
      type: object
      properties:
        id:
          type: integer
          example: 1001
        room_id:
          type: integer
          example: 42
        user_id:
          type: integer
          example: 17
        check_in:
          type: string
          format: date
          example: "2026-06-01"
        check_out:
          type: string
          format: date
          example: "2026-06-05"
        guest_name:
          type: string
          example: "Jane Doe"
        guest_email:
          type: string
          format: email
          example: "jane@example.com"
        number_of_guests:
          type: integer
          nullable: true
          minimum: 1
          example: 2
        special_requests:
          type: string
          nullable: true
          maxLength: 2000
          example: "Late arrival after 22:00"
        status:
          type: string
          enum: [pending, confirmed, refund_pending, cancelled, refund_failed]
          example: "confirmed"
        amount:
          type: number
          format: decimal
          nullable: true
          example: 480.00
        payment_intent_id:
          type: string
          nullable: true
        refund_status:
          type: string
          nullable: true
          enum: [pending, succeeded, failed]
        refund_amount:
          type: integer
          nullable: true
          description: "Refund amount in cents"
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        cancelled_by:
          type: integer
          nullable: true
        room:
          $ref: "#/components/schemas/Room"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Room not found"
...
paths:
...
  /v1/bookings/{booking}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel booking with refund
      operationId: cancelBooking
      description: |
        Cancels booking with refund calculation:
        - **Full refund**: ≥ 48 hours before check-in
        - **50% refund**: ≥ 24 hours before check-in
        - **No refund**: < 24 hours before check-in

        Error codes surfaced in the `error` field of non-200 responses:
          - `unauthorized` — actor is not the booking owner and not admin.
          - `not_cancellable` — booking is in a terminal state.
          - `already_started` — booking check-in has passed.
          - `refund_failed` — Stripe refund raised; `retryable` indicates
             whether the operation can be retried safely.
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - in: path
          name: booking
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled (with refund info)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Booking"
        "403":
          description: Booking cannot be cancelled by this actor (ownership / policy)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                  error:
                    type: string
                    enum: [unauthorized, already_started]
                    example: unauthorized
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Booking is not in a cancellable state
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                  error:
                    type: string
                    enum: [not_cancellable, already_started]
                    example: not_cancellable
        "500":
          description: Refund processing failed (Stripe API error)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                  error:
                    type: string
                    example: refund_failed
                  retryable:
                    type: boolean
                    description: Whether the refund can be safely retried
                    example: true

  # ═══════════════════════════════════════
  # ADMIN BOOKINGS (v1)
  # ═══════════════════════════════════════
...